// src/types/employeetypes.ts
// In your types file
export interface Employee {
    id: string;
    name: string;
    Snapshot: string;
    designation: string;
    timeStamp: number;
    purpose: string;
    department: string;
    borderColor: string;
    age: number;  // Add this line
    remark: boolean;
    image: string;
    type: string;
    role: string;
    ProfilePic: string;
    PhoneNumber: string;
    Person_id: string;
    Vehicle: string;
    VehicleNumber: string;
  }
  